Bug Fix: java_ext folder may not accessable if system umask is too restricted.
Based on chef document(http://docs.getchef.com/resource_directory.html#recursive-directories), when create directory resource rescurively, only the leaf directory will be set to the mode specified in the resource delcaration, all non-leaf directories that is created will have default mode based on the system umask.

On system with umask other than 022, the parent folder of jce may not have executable bit set, which makes the folder not listable for non-root users. This causes the error for example when tomcat server is start with tomcat user, it is not able to load the jce jar files and end with 'unable to find local policy' exception.

To fix the issue, we could use bash to create directory with a umask value set, insetad of using directory resoruce.